Fix Android exception jni type (#29066)

+++ src/app/server/java/src/chip/appserver/ChipAppServerException.java
@@ -3,11 +3,11 @@ package chip.appserver;
 public class ChipAppServerException extends RuntimeException {
   private static final long serialVersionUID = 1L;
 
-  public int errorCode;
+  public long errorCode;
 
   public ChipAppServerException() {}
 
-  public ChipAppServerException(int errorCode, String message) {
+  public ChipAppServerException(long errorCode, String message) {
     super(message != null ? message : String.format("Error Code %d", errorCode));
     this.errorCode = errorCode;
   }
